Starting Puppets & the Republic of Venice
Venice has been added to the 1789 scenario. It starts with cores on all
of its held territories except for Lombardia. The game is set to remove
these cores before the 1804 scenario begins to reflect the change in
thought about the Republic's holdings. The starting puppets have been
modified to be in-line with the 1789 start date as well.